[factory] Add factory store providers component.
Adds component that implements all discoverable fuchsia.factory protocols. This
component transforms a bootfs payload fetched from fuchsia.boot.FactoryItems
protocol to the appropriate output for consumers for factory components.

Files to expose are defined in a set of config files that must appear
in the component's package or config data. In a followup CL, validation
capabilities will be added to verify the contents and length of the
factory files before exposing them to clients.

Moves //garnet/bin/factory to //src/factory.

Test: fx set core.astro --with garnet/bin/factory/factory_store_providers:tests
      fx ota
      fx run-test factory_store_providers_tests
Bug: 33403
Change-Id: I1318f3f9d9b4dce0f90ab6b01e02e84d74392d31