[zircon][sched] fix an untimely yield schedule
The 'yield' operation is done by setting the time slice of the
current thread to the difference between the current time and the
start time. If the thread's relative weight in current CPU has
changed before the 'yield' operation is performed, the current
thread is set to an unexpired time slice. It will not yield
immediately.

Change-Id: I5de29cdab60d9185dfb1b4460ebea74495a08e3a