[componentmgr] Core `Actions` implementation
This change implements support for "actions", which represent a kind of
long-running operation in component manager that can be initiated from
different places. See actions.rs for details.

This change doesn't implement the actions themselves, only the API to
use them.

CF-805 #comment Core actions support, to be used for component
destruction

Change-Id: I09ad45bcc6a0c5b316d99f0e8328cd18016172e7