[syscalls][kazoo] Add various alias workarounds
This is part 2 of a broken up
https://fuchsia-review.googlesource.com/c/fuchsia/+/311026 (which adds
the first backend that uses type information).

This CL adds support for special handling of aliases. In fidl these look
like:

  using something_magic = int;

and kazoo knows that "something_magic" is special. This is used for
various purposes, where the fidl frontend doesn't yet have the
syntax to express the things the syscalls need. For example:

  using mutable_vector_paddr = vector<paddr>;

In the future this would be directly expressed as
mutable<vector<paddr>>. Similarly, names like optional_xyz to indicate
optional out parameters, or vector_xyz_u32size to indicate a vector of
xyz's where the size is a uint32_t rather than a size_t.

A basic version of loading struct definitions is added too, because some
of the aliases contain structs so they need to be resolved.

Additionally, because it's helpful for testing purposes this adds the
GetCName() utility function which maps from Type() to a C string that
the kernel headers would emit. This is useful to confirm that the fidl
type is getting mapped to the right C types/names. New tests are added
for the alias mappings.

Bug: 33668
Change-Id: I0eecf101690ee6e14a93c0556509f7fe6c5cd835