[fidlc] typeshape -> typeshape()
This CL changes the public |typeshape| member variable on a few structs
to be a private |typeshape_| member variable. We then add public
typeshape() and mutable_typeshape() accessors to access the object's
typeshape.

While this is a simple, mechanical change, the intent is for objects to
properly encapsulate how their typeshape is accessed, so they can change
their internal implementations later.

Test: fx build zircon/system/utest:host && out/default.zircon/host-x64-linux-asan/obj/system/utest/fidl-compiler/fidl-compiler-test.debug
Test: fx run-host-tests fidlgen_cpp_test fidlgen_golang_test  # ensure fidlgen tests pass
Test: fx set bringup.x64 --with-base //garnet/packages/tests:zircon --variant asan --variant host_asan && fx build && fx run -k -c zircon.autorun.boot=/boot/bin/runtests+-t+fidl-coding-tables-test  # ensure coding tables test pass

Change-Id: I74142cbf62bd276fc158384d7575627e1cc94467