[dhcpd] Implement response routing behavior per RFC2131
This change implements RFC2131, Section 4.1, Page 22, Paragraph 4,
which specifies how to define destination addresses for server responses,
and updates the code and tests accordingly.
In also adds missing implementation in the server wherein it must set broadcast
bit in the nak if the relay agent ip (giaddr) is set. It also gets rid of
some hardcoded zeroed ipv4 addresses.

Change-Id: I5d46d1a4b18a6a746bb96eedd5824864ee925fcd