[debugger] DebuggedJob.SetFilters returns koids of matches.
Before the debuged job would immediatelly attach to each process.
Now it also logs the matched processes, so that it can communicate them
back to the client.

This is part of a broader scheme of having the client being more
explicit about process handling by being able to query the current state
and having the client decide what to do.

This attaching part is still there, so this should have no logic change.

TEST=Added unit test to verify correct process filtering/attachment.

Change-Id: If5797d9cc6605f60a6d94b053f16c08564ff15ae