[kernel][syscall] Limit len to vmo_read/write to prevent overflows.
Limit the len to vmo_read and vmo_write to prevent overflows, that could
potentially cause infinite loops. (This is an unlikely but theoretically
possible bug). This check (along with the loops that cause the
pre-faulting) are part of a workaround which will be removed when the
underlying bug is fixed.

Bug: ZX-4642.
Test: Added vmo unit test that tests that the expected error is
returned for large invalid len values.

Change-Id: I2f51040a0ddff4d163318c808dab785d1154a399