[acpi][i2c-hid] Workarounds for Slate touchscreen
The Slate touchscreen needs to be explicitly powered on via ACPI when
the device starts up. Currently we do not have a rationalized way to
manage power states of ACPI devices, so a temporary workaround is
necessary.

Installs a temporary workaround during ACPI tree traversal (which
happens on device startup) to detect the hardware configuration used on
Slate and send the power-on signal to the touchscreen. A similar
workaround already exists for Acer laptops.

Also makes the i2c-hid driver retry the first few i2c transactions three
times, since there is a delay between turning on the touchscreen device
and it being responsive on the i2c bus. Ideally a principled mechanism
for checking the device's readiness would be used, but there is no such
mechanism I am aware of (ACPI _STA returns a hard-coded value, for
example; it does not indicate actual device readiness).

Bug: 34651

Tested: Added device-enumeration test for Slate; manually tested.

Change-Id: I238a7c3cb4d713f30bad6a44fae25e35c49804a7