[kernel][vm] More efficient page attribution
Rewrites the logic for determining ancestor page attribution to operate
on ranges instead of probing every potential page. This reduces the
complexity from O(sd), where s is the size of the vmo in pages, to
O(nd), where n is the number of the committed pages in the vmo
hierarchy and d is the depth of the vmo hierarchy.

In my testing this gives an order of magnitude performance improvement
in querying VMO information.

Bug: 34652

Test: The vmo-clone tests in the core tests comprehensively test that
page attribution is correct.

Change-Id: I3ebe4c97b164cae75ef8074b36d5787060daea22