[syscalls][kazoo] Use aliases in existing syscall definitions
This removes a variety of unused attributes (the 'In0="something"' style
ones) in favour of the aliases for which support was added in a previous
CL.

Additionally, it removes the similar attributes that were used for
static analysis (which wasn't working in abigen/banjo anyway and has
been removed). These look lik handle:acquire handle:release_always, etc.

Bug: 33668
Change-Id: I5b42d02d0519c50ae6ca629906333294eb5edcc9