[cobalt] FuchsiaSystemClock implementation of the ValidatedClockInterface in Fuchsia.
The new FuchsiaSystemClock is initialized by the Cobalt app, but the callback
when the clock becomes accurate is currently a no-op. Once the other parts of
the inaccurate clock design are implemented, this callback can be used to start
the event aggregator and flush the saved events from the UndatedEventLogger.

Change-Id: I84a4da38a737f1c2f81649ccc10e20072cfc6779