[Netstack] Cleanup when DHCP loses its lease or stops running
Note that due to b/138755518 the removed IP address may remain on the
interface until all sockets that have an active connection with this
address are closed.

This change also tweaks the route table logging and cleans up the use
of '%v' in routes.

Bug: 21528
Tests: Manual tests (may need to be run via console)
  ifconfig ethp001f6
    (observe interface has address)
  ifconfig ethp001f6 down && ifconfig ethp001f6
    (observe address was removed)
  ifconfig ethp001f6 up && ifconfig ethp001f6
    (observe interface has same address again)

Change-Id: I028ba94ed33810831ae3c62d93741c57558f8e2f