[cobalt] Stop basing the debug/non-debug decision on the channel.
Previously we were using the channel, as specified by Amber, to
decide whether or not the currently running system should allow
debug-level metrics to be collected.

We replace that with a new function DetermineReleaseStage().

Currently DetermineReleaseStage() has a dummy stub implementation that
always returns DEBUG. This is sufficient for the next month.

When Fuchsia develops a mechanism to make this determination, we
will reimplement DetermineReleaseStage() to make use of this
mechanism.

Bug: 34960
Bug: 3886
Bug: 3826
Change-Id: I5eefaac1e6d1f11a1fee2c1fb4f16c3c6b0a0f01