[ledger] Fix emptiness in SetActivePageManager
By moving the assignment of the instance fields to late in the body of
SetActivePageManager we make the ActivePageManagerContainer unable to
become empty during the calling of the internal request callbacks.

We also allocate all ExpiringTokens to be passed to internal request
callbacks before calling any internal request callback to avoid the
case of the tokens dithering between zero and one and having the
TokenManager's on-empty callback called more than once during the
calling of the internal request callbacks.

The content of this commit probably should have been included in
655ff3971df994355f5dcac82ed9c071b6ed53d5.

A TODO linking to
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=35152 which
describes a potentiality for making ActivePageManagerContainer's
internal state more robust is also included.

Change-Id: If7156f413162265df9d317efe2f8b8c964ff5866