[syscalls][kazoo] Implement vdso-header backend
With the previous CLs in place, outputting the vdso header is fairly
straightforward.

To reduce number of CLs by one, in addition to the straightforward
vdsoheader output printer, this CL also adds support for loading enums,
and for a fidl-level attribute named ArgReorder which allows reordering
the kernel arguments to match existing convention.

Tests of vdsoheader are also added, and the output (via difftest) is
identical to the existing output from abigen.

Bug: 33668
Change-Id: I58f8fae826832e58d251cd21dd2ea2d18898b551