[debugger] Debug logging is now a tree.
Linear output was getting hard to read.
This change makes every logging statement makes a push on constructor and pop
on destruction (end of block). This permits to nest messages that are
done deeper down the stack and being able to show them indented, thus
showing that a message was done within the context of another.

TEST=Manual testing over a representative case (somewhat involved).

     On a shell: fx shell -- run debug_agent.cmx --debug-mode
     Another shell: fx debug

     Have an assert somewhere in the code (I used debugged_thread.cc)
     and watch the result.

Change-Id: Ie0cf8d419f5ec3ddab7e2baba75eed9f6aca13ad