[audio_core] output glitch log tuning
This CL contains two small changes made while debugging a timing
and buffering issue with the audio capturer. First, it adds trace-
logging for fifo_depth.  Secondly, it clarifies the mix-underflow
error message, and makes the "recovered from underflow" companion
message a WARNING (not INFO), so logging will include not only the
time that the underflow began, but also the time by when the
system had fully recovered.

Test: logging-only

Change-Id: I80bc84654867941040d970c6e3a02e537f0bdd08