[scripts][fuzzing] Fix logging and symbolization
This CL replaces `fx fuzz`s use of loglistener with a dlog-based
approach: the code first parses the libFuzzer logs to looking for PIDs
associated with errors, then uses dlog to retrieve just the stack traces
for those processes. These stack traces are symbolized, and the crash
trampoline frames are truncated, if present. The resulting stack trace
is then inserted in-line into the libFuzzer log to mimic those produced
on Linux as closely as possible.

Bug: SEC-352
Test: fx run-host-tests host_fuzzers_test
Change-Id: I1704d32626cc5acbc087135637d5d837c620a17f