[hid] Convert to static library
The hid library does not need to be a
shared library. There are a few functions
which use shared library features but these
functions are no longer used.

Convert the hid library to be shared and
remove the temporary flags on each of the
drivers that use the HID library.

TEST: Compiling correctly is the main test
  here, and CQ should catch any regressions
  in any of the individual drivers. This
  is unlikely to cause an error.

Change-Id: I0e5b8dccfa96b24b37dcd40008e6ef5ed1fcba3c