[prebuilt_tests] Passing the prefixed-package name to run.
Currently, the topaz-x64-debug-chromium_prebuilts builder is failing during test execution due
to the fact that the packages that are baked into the system image do not match the package that's
invoked via the "run" cmd.

This CL renames a variable used for identifying the prefixed_package_name, and passes that variable
to be executed in the "run" cmd.

Change-Id: Ie9ae0127cefff2644f622b59c9d82e5ee2de538d