[utest][threads] Avoid library calls in ThreadStarter test functions
The threads started in the test by the ThreadStarter class do not
provide the full ABI setup.  So code running in those threads must
be restricted to code compiled with the minimal ABI, which includes
only thread-functions.cc functions and the vDSO.  Even "header-only"
functions from libc++ and libzx are incompatible with this execution
environment.

Bug: 3770
Change-Id: Ia1857545378de3134c631745094a1f95ab3c58a2