[pseudo-fs-mt] Expose watcher controllers
Lazy directory will need to construct the EXISTING event sent to the
newly connected watcher using async operations.  Making
`watchers::Watcher:add` async (along with the event producers) seems
like a bit too heavy.

As an alternative, we just expose the watcher controller and make it a
responsibility of the code that is adding the watcher to call
`send_event` to send both EXISTING and IDLE events.

This allows the lazy directory to construct multiple EXISTING events
(each filling the whole buffer) and send them from an async block, later
followed by an IDLE event.

Change-Id: Ib66cd6b7e197355b5a92c3f395a803c344ad5654