[lz4] Re-Add LZ4 Fuzzers
This CL adds back the LZ4 fuzzers, which were removed in CL 305669.

I added fuzzers for lz4 back in CL 225293. At that point, the fuzzer
quickly found some issues and we created a bug to update lz4 to a recent
version. Time passed and the fuzzers got stale during the great
gnification of the zircon repository. Finally, they ended up being
removed in when zircon updated lz4 and moved to a "verbatim" copy (so,
no forked source files).

NOTE that I am adding the fuzzer code here in this folder (which I think
makes sense which is where the library is defined), but I needed to add
a build rule inside system/utest so these fuzzers were included in the
build.

BUG: 35199
Change-Id: If54d66f18b611ab97fc17f185d01ffb24eaf484b