[pseudo-fs-mt] test_utils: Config with builder pattern
There is a number of parameters that can be specified when running a
pseudo-fs test.  Providing multiple different functions to cover all
possible combinations becomes inconvenient - a builder pattern works
better for this kind of task.

On top of that, this change combined
`file::test_util::run_server_client` and
`directory::test_util::run_server_client` backing code.  They were only
different in the type of the proxy that a client test code operates
upon.  And the only reason it was not generalized over that type, is
that the compiler can not infer type correctly in the presence of this
generalization.  But as the amount of code is growing, what used to be a
small duplication became a duplication of almost 300 lines of code.

The above was fixed by providing thin wrappers with explicit types that
rely on the same backing code otherwise.

Change-Id: I6b116e5a192c5d8fb7fcdd69e84a08f26d82c0cf