[wlan-hw-sim] Make connect more tolerant with RSN
Even though RSN exchange could time out, the packets are never lost in
the simulation. With the authenticator and supplicant's ability to
handle replays, the client would always be able to connect successfully.

As long as RSN related error does not cause the test program to panic,
the test would complete successful.

Bug: 35329

MULTIPLY: `[
  {
    "target":
      "//src/connectivity/wlan/testing/hw-sim/test:connect_to_wpa2_network",
    "total_runs": 200
  }
]`

Test: Test only change. Does not affect behavior.
Change-Id: I6e9b08af45be90e03e30823fabf2a1a390155eeb