[componentmgr] Implement per-event-type hooks
Prior to this CL, every Component Manager hook lived on a single trait.

This meant that any object interested in hooking into some component
manager event received all events and was forced to stub out events
that it was not interested in.

This CL splits off hooks into per-event traits and pulls hooks into
a thread-safe cloneable object "Hooks" enabling hooks to be installed
at runtime. This will likely be useful in the future when we wish
to support Archivist and debugger hooks in the component manager.

CF-541
Change-Id: I4d090d7c81a3031df66248dc99644cbe2bbace4d