[build] Build the vDSO in a bespoke compilation environment
The vDSO has very special build requirements.  Heretofore these were
all satisfied locally by the setup of the targets in the vDSO's own
BUILD.gn file.  However, the vDSO will soon start to use code in its
implementation from other libraries that are also built in more
normal circumstances.  For this to work, all that other library code
also needs to be compiled specially for the vDSO.  That's exactly
what environment() is for, so use it now.

Bug: BLD-583 #done
Change-Id: Ib0942c17899b15004b071f33d18ced6c2e4dabf4