[netstack3] Enable and disable interface when driver changes state
Ethernet status changes received by the bindings now modify interface
status correctly. This results in the interface being removed from the
core with all pending timers deinitialized, or added back to the core
and re initialized in the bindings.

Tests: Added test that toggles the netemul device driver and ensures
that the reported end to end state is reflected properly in the
bindings.

Change-Id: Icecd8409a7b9506c3b92feb9ed46a4d27df0203d