[pc][mexec] Double mexec identity map size.
The perf-dawson_canyon tests were failing because mexec was unable to
allocate 700MiB of contiguous RAM below the 4GiB mark.

This was due to a discontinuity in the device's physical address space.
To remedy this, we increase the size of the identity mapping to 8GiB
which is large enough to accomodate the incoming image.

Change-Id: I4e264fe16e5cf190f1dee2f1d3bd87d784410993