[syscalls] Change zx_debug_read() signature
debug_read is the only syscall that currently uses an in/out for the
size of the buffer. Normalize this by making it a buffer+size and
actual, similar to other syscalls. As there appears to only be one
callsite, this is a relatively straightforward change, despite it
changing the public API.

(The side benefit of this is that it allows the upcoming fidl-expression
of syscalls to use a string type for the buffer, and doesn't need
special handling to make length of that buffer mutable.)

Bug: 30517
Bug: 33668
Change-Id: I365292d6219534a2ade59b7ba7a6d10d28af2d8a
Exempt-From-Owner-Approval: rust owners appear to be OOO at rustconf, and the .rs change is mechanical, so I think a post-review is probably OK