[pc][mexec] Use 64-bit instructions to load the bootdata address.
Mexec may place the bootdata above 4GiB which means that the bootdata
address will be 64B wide. Using 32 bit instructions was lopping off the
high half of the bootdata address.

Note that elsewhere in the function we use 32bit instructions because
the addresses and offsets they refer to are actually 32 bits wide.

Change-Id: Iad8a303a8ffe23ea307bdce18b57d5a5f38a8f7d