[mexec] Enable mexec to memmov multiple regions of memory.
Previously mexec was only capable of copying one region of memory
(the kernel). This patch enables the mexec syscall to relocate a
vector list of regions to relocate before branching to the new kernel.

This allows us to move the bootimage to arbitrary locations to test
that we are capable of booting the system regardless of where in RAM
we place the bootdata.

Change-Id: I6457ea61160f64dd4c496596846c53353473cdae