[fidlc] Fix tables for foreign nullable xunions
This CL fixes an issue with how the coded types generator compiles
foreign coded xunions. Previously, when a library used a nullable xunion
from a foreign library dependency, a new nullable coded type would be
created but its fields would not be resolved. This causes an incorrect
table to be generated, resulting in UB at runtime since the walker does
not know how to properly visit and validate the aforementioned type.

This change resolves the issue by emitting both the non-nullable and
nullable xunion tables in the libraries in which the xunion was defined.

Added new tests to fidl-compiler-test and fidl-coding-tables to catch
this.

Bug: 35210

Test: fx build zircon/system/utest:host && out/default.zircon/host-x64-linux/obj/system/utest/
fidl-compiler/fidl-compiler-test
Test: fx run -k -c zircon.autorun.boot=/boot/bin/runtests+-t+fidl-coding-tables-test

Change-Id: I543b1e876d4f9087363adc103053eab932d778ad