[cobalt] Removes call to soon-to-be-removed function.
We plan to remove the method ProjectContextFactory::is_single_legacy_project()
fromm Cobalt Core because we plan to remove legacy logic for parsing
a legacy project. In this CL we remove a reference to that function.

There is no need to special-case seeing the bytes of a legacy project.
In the future that will be handled in the same way as any other sequence
of bytes that does not constitute a valid project in the current model.

Change-Id: I1b4cff81cbb62d75814f488164c37df2c281d2f6