[serve-updates] avoid process spawn in live checks
There exists a bug elsewhere in the stack that sometimes causes process
spawning to take an exceptionally long time. This in turn causes the echo
based live check to lock the sshd process long enough that it misses enough
keepalive packets that the client terminates the connection.

We hereby adjust the mechanism used as a live check to instead just poll the
ssh client control master process for liveness, avoiding the process spawn.
This does not prevent the issue, as other program invocations for other
purposes can still trigger the conditions, however, this removes the most
common cause of the user reported symptoms that are leading to confused
diagnosis of the root cause.

This also should essentially remove the more scary symptom, wherein sshd
exits before the echo process has actually spawned, resulting in
ZX_ERR_PEER_CLOSED on the ldsvc channel.

Test: serve-updates "survives" while k ut -r 10000 all is running
Bug: 35333
Bug: 33470
Change-Id: I5c7c43fa1b51a408fc8772299f60048a826a03fc