[debugger] Debug agent sends matching pids on filter request.
Before it would just send the zx_status_t. With this the client will
know explicitly which processes matched. This is part of a larger change
in which the client will explicitly ask for those processes to attach
instead of thema attaching implicitly. This gives more control to the
client and permits to reduce the amount of attaching flows within the
agent.

On the client side, the job context will now issue a filter match event
on to the observers. No one is currently listening to this event, so
there should not be any logic changes.

TEST=Added JobContextImplTest.
     fx run-host-tests zxdb_tests

Change-Id: I06ad262b2395265d92a91c7ae124a0fbec9dbe9c