[build] Remove bogus tools targets from //zircon/BUILD.gn
The reorganization of //zircon/BUILD.gn and the move of the tools
build logic into //zircon/tools/BUILD.gn were ships passing in
the night and getting turned around by each other's wake.

Note that the top-level "tools.all" target is no longer provided.
The Ninja target for that is "tools:all-hosts".

Bug: 35313
Change-Id: I795685fab3377b50414b28c22b9076787168d178