[debugger] Add several log statements to debug agent.
While debugging fidlcat, several log statements have proven useful.

Added a DEBUG_LOG_WITH_LOCATION so an explicit location can be passed in
from the macro instead of using the implicit FROM_HERE. This permits to
a function to delegate the printing to another function without having
it declare a wrong log location (the printing functions instead of
issuing one). This permits less clutter in the actual logic.

TEST=These are all logging statements. There is no logical change.

Change-Id: I8deeb31f6bd82002442a19bf3e89ca4bd6ba1906