[x86] Add local tracing support for idle loop
Adds trace duration events tracking how long the CPU was idle in the
idle loop, and the MWAIT hint that was passed in as an argument to
idling (i.e. which C-state was requested).

Disabled by default since this results in a high volume of ktrace
records.

Example trace is provided in the linked bug.

Bug: 35289
Testability: Tracing-only change
Change-Id: Ia397a6935a55384366265856e54455084b55f2cc