[ledger] Write entry_ids in treenodes
The entry ids given to btree::ApplyChanges are written to the nodes (for
insertions and updates). The journal now generates the dummy entry ids
that were previously generated on reading (but this code is kept for
compatibility with nodes written without entry ids).
Adds a test for journals checking that identical merges generate the same
commit.

Co-authored-by: Nelly Vouzoukidou <nellyv@google.com>
Change-Id: I25c114c325882e4973eb5216d649d5ef9a1904e1