[netstack] Make DNS failure logs debug verbosity
Often, client applications will make failing DNS calls in a loop, which
causes logspam. Typically in these cases, the problem is an underlying
networking issue (or simply running on a device that is not connected to
the network), so the DNS logs are a red herring for folks who are
actually trying to debug the problem.

Change-Id: Ia0e977563ff2a56e9ba97c86b86d9973f5e67029