[fidl-utils] update usage comments
This change expands the usage comments for the FIDL BindMember util to
include examples of non-returning FIDL methods. This will hopefully
prevent the need for future authors to spend time working backwards from
obscure 'attempt to use a deleted function' template compilation errors
when using the utility with incorrect signatures.

Change-Id: Iabb55e42e79b1b77b52bfaa53fdbf5519411886a