[devhost] Make device_rebind work when a device has no children
The current behavior was forcing us to manually bind the GPT driver in
the paver, which leads to buggy/racy behavior. By fixing rebind
behavior, fshost is now able to observe the rebind event, and bind the
GPT driver automatically.

BUG: 35274

TESTED: Initialized a device with no valid gpt partitions into one with
a valid GPT header and a single partition and observed add/remove events
in fshost.
BUG: runtests -t devhost-test-test -- --gtest_filter=CoreTest.*

Change-Id: Ib3e1f697dc3a096637bc69b5de8ac0df574fa627