[paver] Stop attempting to manually bind GPT driver
Rebind behavior was fixed to allows fshost to always see the rebind
event, so that it could auto-bind the GPT driver. This allows us to
avoid racy behavior that occurs when two entities try to bind a driver
to a device concurrently.

BUG: 35274

TESTED: Paved a device with no valid GPT partitions successfully.
TESTED: runtests -t paver-test (existing tests catch the issue when
        block-watcher is enabled)
Change-Id: I6bd38fac1ef462b2423463d40d3727665b7d7d41