[discovermgr] First step towards s/ModuleOutputWriter/StoryModule
This brings the new StoryModule interface and the implementation of the
equivalent of ModuleOutputWriter#Write as WriteOutput. StoryModule
interface will also implement AddModuleToStory and provide controllers
to clients.

After the single user of ModuleOutputWriter is migrated to the new API,
I'll go ahead and remove it.

TESTED=fx run-test discovermgr_tests -- --nocapture

Change-Id: I46f7f2330e62faadf63ea1a86dbbe7f083a3c2c6