[kernel] Complete the implementation of channel_write_etc syscall
This channel write variant allows allows one to specify the disposition
of the outgoing handles, in particular it allows to:

1- duplicate and send with reduced rights
2- move and send with reduced rights
3- verify handle type before sending

The duplicate option saves one or more syscalls, for example before:

    zx_handle_duplicate(event, ZX_RIGHT_TRANSFER | ZX_RIGHT_WAIT, &new_handle);
    zx_channel_write(channel, ... &new_handle, 1, .. );

The move option can also reduce syscalls.

Both options allow removing ZX_RIGHT_TRANSFER, which was not possible before.

Test: Added core/channel-write-etc-test which passes.

Bug: 34144

Change-Id: Icd4805b48507c44e996c4cf9ea8cfa5369488212