[devcoord] Fix unbind task race condition.
We do not want to add an already posted task as a dependent,
otherwise when we complete we will try to post the task again.

This can happen if DdkRemove is called on a parent device and
child device at the same time.

Example
-----------------
Parent Device A
Child Device B

DdkRemove(A) -> UnbindTask(A) posted
DdkRemove(B) -> UnbindTask(B) posted

Async loop runs UnbindTask(A)
UnbindTask(A) adds UnbindTask(B) as a dependent
UnbindTask(A) completes immediately
UnbindTask(A).Complete tries to post UnbindTask(B)

BUG=35361

Change-Id: I376bceb219601088c7345a0169041c27770a42f7