[go/inspect] Inspect VMO buddy allocator and blocks
This change implements the lower levels of the Component Inspect VMO for
Go -- the block API and buddy allocator. This is a preliminary
implementation that may change as further work on the public API informs
changes.

The block API is generated using `go generate` from a template; Go's
lack of tagged unions, polymorphism, generics, and bitfields all make
this a bit clunkier than it otherwise might be.

Change-Id: Ie4a5621277b36c24bcc7adbfe880a1923c08c734