[kernel][vm] Avoid unmapping on COW child create
This optimizes the COW child creation process to minimize the required
mapping changes. All that is required is that anything that was
writable ceases to be writable, this means
 * Read only or execute only mappings do not need to be removed.
 * Writable mappings can be downgraded to read only.

The current implementation in VmMapping will not actually downgrade
mappings, but will only skip read only mappings or unmap. This still
provides a significant optimization for read only mappings, such as
those used by code, to not be unmapped.

Importantly this only provides an improvement for VmMappings that are
readable. If the VmMapping is writable, even if the page is only mapped
into the hardware page table as read only, it will still be unmapped.
This is a limitation in the ability to bulk query a range of mappings
from the address space.

Change-Id: I304f07452034b654dc572cfc14504884e988a0a9