[ISP] Setting Isp interrupt to level mode
The interrupts were not getting triggered when we were running the ISP.
Setting the pulse_mode to level causes all the interrupts to be caught.

The board file is already set to trigger on rising edge, so it is unclear why
this fixes things.  More investigation may be needed.

CAM-129 #done

Testing:  This fixes a broken test: camera_full_on_device_test
For information on how to run this test, see //src/camera/TESTING.md

Change-Id: I0d47fb5a211e93c6884216c2ed287b6040e08214