[audio_core] Simplify sending SendPacket/DiscardAllPackets callbacks.
The problem being solved here is getting the FIDL callbacks from
SendPacket or DiscardAllPackets invoked on the main FIDL thread after
the requisite processing has been done.

Previously this was done as follows:
  * Once every link has finished processing the packet, the
  AudioPacketRef will be popped from the link's queue.
  * After the refcount goes to 0, the fbl_recycle method will be called
  on AudioPacketRef.
  * AudioPacketRef sees it's being recycled for the first time and it
  has a callback. It invokes AudioCoreImpl::SchedulePacketCleanup.
  * The AudioPacketRef is inserted into a linked list and a task is
  posted to the main thread to drain that list.
  * The task on the main thread takes all the packets and invokes their
  callbacks before dropping the AudioPacketRef.
  * fbl_recycle is called for a second time on AudioPacketRef. This
  time the AudioPacketRef see's it's already been recycled once so it
  deletes itself.

The key insight is that after the first fbl_recycle, no part of the
packet is referenced anymore except for the callback. Since the
async::Loop already is maintaining an ordered list of tasks, we
short-circuit by just posting the callback directly to the async loop
and removing the intermediate data structures.

Change-Id: Ibef46e8678c0e9ed4f51311adb1f40b93e71def4