[netsvc] Don't require --advertise to start netsvc server.
This restores the --nodename flag to netsvc (removed in 7ea67b9) to
print the device's nodename and exit immediately, this is used by
Infra to extract information about the device from channels outside
of the netsvc e.g. mapping serial to the correct Fuchsia device.

Similarly, Infra does not pass the --advertise flag in most common cases
because it is believed to cause too much broadcast traffic on local
networks with many Fuchsia devices connected.

Change-Id: I804961f413611d1220be1882820c4a3fda3b4711