[rust][inspect] Better support for array reading by taking format into account.
This will be handy on iquery. It's also handy for testing given that the
way components write arrays is by specifying the format, floor, step size, etc.
The testing module still forces to understand how the values are
stored in the vector, so there's area for improvement there in a follow
up.

TESTED=fx run-test rust-crates-tests -t fuchsia_inspect_lib_test -- --nocapture

Change-Id: I40d0bcb4b863038a65e60c905c8786cf92a2a168