[fidlc] Remove typeshape member from FieldShape.
The FieldShape struct currently contains a TypeShape. This is sometimes
more convenient, since struct members can e.g. just contain a FieldShape
instead of both a FieldShape and a TypeShape, but it means that
computing the FieldShape independently of TypeShape is more challenging,
since both have to be computed. In short: FieldShape complects
TypeShape[1].

There are also parts of the code where FieldShape is used unnecessarily
instead of just a TypeShape. Separating the two types makes the
distinction more clear.

The static {Struct,Union,etc}::Shape() method APIs become more
cumbersome with this CL, but the intention is for this CL to be
short-lived. The next CLs in this CL chain will compute the fieldshape
& typeshape on-demand, so that TypeShapeContainer, FieldShapeContainer,
and all the {Struct,Union,etc}::Shape() methods can go away. \o/

1. With thanks to Rich Hickey, where I learned "complect":
   <https://www.infoq.com/presentations/Simple-Made-Easy/>. See 31m36s.

Test: fx build zircon/system/utest:host && out/default.zircon/host-x64-linux-asan/obj/system/utest/fidl-compiler/fidl-compiler-test.debug
Test: fx run-host-tests fidlgen_cpp_test fidlgen_golang_test  # ensure fidlgen tests pass
Test: fx set bringup.x64 --with-base //garnet/packages/tests:zircon --variant asan --variant host_asan && fx build && fx run -k -c zircon.autorun.boot=/boot/bin/runtests+-t+fidl-coding-tables-test  # ensure coding tables test pass

Change-Id: I66a63c876b5500c5d4f15ffcf97c1a92626c4774