[build] Allow FIDL libraries in Zircon to be addressed by their real path.
The BUILD.gn file for these libraries now works in both builds!

The idea is to bypass the GGN mechanism in //zircon/public - which still
exists in order to softly transition existing references to libraries - and
directly depend on library via its build file.
Once the ZN build does not depend on a FIDL library anymore, that library can be
freely moved to another part of fuchsia.git (e.g. //sdk).

Bug: 3367
Change-Id: I8ab90c256cc4f30414309b0081d8553fd51a9042