[mediaplayer] support sysmem payload management
This CL adds sysmem support to mediaplayer. Many files are touched,
but the focus of changes is the PayloadManager class, which now has
extensive tests.

The fidl decoder (the wrapper for mediacodec) doesn't yet use the
sysmem features of the payload manager. This will be added in a
subsequent CL.

Support for physically-contiguous buffers is temporarily broken. The
CL that brings the fidl decoder up to date will address this via
sysmem.

TEST: new payload manager tests
      existing CQ tests cover any other possible regressions

Change-Id: I497e67cdfc44f6ef899d0d38eb2b01e9fcd39e51