[hidbus] Update GetDescriptor to not allocate
The GetDescriptor banjo protocol would rely on
the HidBus driver to allocate the descriptor
and then pass it to the higher driver.

This leads to a host of memory problems in
the HID_DEVICE driver where it attempts
to free the descriptor that may have been
allocated from several different places
in different Hidbus drivers.

Update the protocol so the higher driver passes
the buffer back to the Hidbus driver.

BUG 35258

Change-Id: I33b0e750c358f4dbc970ac552530287e985ef176