[componentmgr] TestHook should dispense its own hooks
TestHooks::hooks(hook.clone()) was a bit painful for me to read in
code review so I decided to simplify it a bit like Hub.

TestHook has an inner: Arc<TestHookInner>, and so TestHook::hooks()
can dispense Arcs to Inner as part of the hooks vector.

This is a minor change but it feels like slightly less of an eyesore.

Change-Id: Id620607931f35254ca66d85f30957d22fb0960b5