[minfs][2/2] Add support for backup superblock.
Reference design doc :  go/backup-superblock-minfs

This change adds support for a backup superblock in minfs. It adds a checksum in the superblock,
and makes sure that the checksum of superblock is correct everytime we update it/check it.
It also adds a "generation_count" in the superblock for debugging purposes.
Since this is a breaking change, we have added a "major version" and "minor version" to safeguard
us from future breaking changes for backward-compatibility. Using "reserved" field at the end of
superblock forcing the superblock size to be same as a block size. This will guard against any
future format changes to the superblock and the reserved fields can be used to append more newer
fields to superblock structure in future.

This change safeguards against bit flip/on-disk corruption of superblock due to the following bugs:
1. Filesystem Software bugs
2. Block device Software bugs
3. Hardware corruption bugs when superblock / backup block.

Fixes: ZX-4820 #done

Misc:
    1. Additionally inspector was updated to show these new fields.
    2. BOOTLOADER_VERSION was updated from "0.7.17" to "0.7.18" since this is a
       breaking change and will need a zedboot update.

TESTING: 1. Added unit tests for superblock to test various corrupt/non-corrupt/FVM/non-FVM cases
            of the superblock.
         2. Added end-to-end integration tests using ramdisk to test backup superblock
            functionality.
         3. Added unit test for testing bitmap reconstruction logic.
         4. Added unit test for testing soft upgrades from version 7 to version 8.
         5. Inspector tests were updated to test the new fields.
         6. Since this is a breaking change, manual testing on major
            hardware we currently support is done to make sure we do not break anything.

Change-Id: Icd7b13aa1f35ce6683fb09ee050b6da84db8d3c0